# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления фильмами и пользователями с аутентификацией",
        "title": "Movie REST API",
        "contact": {
            "name": "API Support",
            "email": "st1txh.devops@ussr.com"
        },
        "license": {
            "name": "Omsk",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/films": {
            "get": {
                "description": "Retrieve a list of all films with optional sorting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Get all films",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field to sort by (title, rating, release_date)",
                        "name": "sort_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of films",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_films.Film"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new film with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Create a new film",
                "parameters": [
                    {
                        "description": "Film data to create",
                        "name": "film",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_films.Film"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created film",
                        "schema": {
                            "$ref": "#/definitions/internal_films.Film"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/films/sorted": {
            "get": {
                "description": "Retrieve a list of films sorted by specified criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Get sorted films list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field to sort by (title, rating, release_date)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order (asc, desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sorted list of films",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_films.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid sort parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/films/user/{uuid}": {
            "get": {
                "description": "Retrieve all films associated with specific user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Get films by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No films found for user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/films/{uuid}": {
            "delete": {
                "description": "Remove a film by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Delete a film",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Film ID (UUID)",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Film deleted successfully"
                    },
                    "404": {
                        "description": "Film not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update specific fields of a film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Partially update film",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Film ID (UUID)",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_films.UpdateFilm"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Film updated successfully"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_films.Film": {
            "description": "Модель фильма с рейтингом и датой выпуска",
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "created_at": {
                    "description": "@format date",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "film_id": {
                    "description": "@format uuid",
                    "type": "string"
                },
                "rating": {
                    "description": "@minimum 0\n@maximum 10",
                    "type": "number"
                },
                "release_date": {
                    "description": "@format date",
                    "type": "string"
                },
                "title": {
                    "description": "@minLength 1\n@maxLength 255",
                    "type": "string"
                },
                "updated_at": {
                    "description": "@format date",
                    "type": "string"
                }
            }
        },
        "internal_films.UpdateFilm": {
            "description": "Модель фильма с необходимым базисом для обновления",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "query"
        }
    }
}